Stack Chart

+++ app/src/main/java/com/example/meezan360/model/footerGraph/data/HorizontalBarChartDataModel.kt
@@ -4,9 +4,9 @@ import com.google.gson.annotations.SerializedName
 
 data class HorizontalBarChartDataModel(
     @SerializedName("key") var key: String,
-    @SerializedName("value") var value: Int,
-    @SerializedName("percentage") var percentage: Int?, //
-    @SerializedName("target") var target: Int,
+    @SerializedName("value") var value: Float,
+    @SerializedName("percentage") var percentage: Float?, //
+    @SerializedName("target") var target: Float,
     @SerializedName("value_color") var valueColor: String,
     @SerializedName("target_color") var targetColor: String
 )
